+++ src/move.js
@@ -2,22 +2,37 @@ import * as MODULE from "../MaterialDeck.js";
 
 export class Move{
     constructor(){
+        this.active = false;
     }
 
     keyPress(settings){
+        console.log('move',settings)
         let dir = settings.dir;
         let mode = settings.mode;
         if (mode == undefined) mode = 0;
         if (dir == undefined) dir = 0;
-        if (settings.mode == '1')
-            this.moveToken(MODULE.selectedTokenId,dir);
-        else    
-            this.moveCanvas(dir);
+        if (dir < 9){
+            if (settings.mode == '1')
+                this.moveToken(MODULE.selectedTokenId,dir);
+            else    
+                this.moveCanvas(dir);
+        }
+        else if (dir == 9) {//zoom in
+            let viewPosition = canvas.scene._viewPosition;
+            viewPosition.scale = viewPosition.scale*1.05;
+            viewPosition.duration = 100;
+            canvas.animatePan(viewPosition);
+        }
+        else if (dir == 10) {//zoom out
+            let viewPosition = canvas.scene._viewPosition;
+            viewPosition.scale = viewPosition.scale*0.95;
+            viewPosition.duration = 100;
+            canvas.animatePan(viewPosition);
+        }
     }
 
     async moveToken(tokenId,dir){
         if (tokenId == undefined) return;
-
         const token = canvas.tokens.children[0].children.find(p => p.id == tokenId);
         const gridSize = canvas.scene.data.grid;
         let x = token.x;