+++ changelog.md
@@ -1,4 +1,22 @@
 # Changelog Material Deck Module
+### v0.8.5 - 17-11-2020
+<ul>
+<li>Added 'Display Icon' to Macro action</li>
+<li>Removed background option from 'Macro' => 'Macro Board' => 'Trigger Macro', this is handled in the macro configuration screen</li>
+<li>Added 'Background' option in 'Macro' => 'Macro Board' => 'Offset'</li>
+<li>Fixed control buttons from not performing action when clicked</li>
+<li>Added check for each function to prevent unnecessary searching through the button buffer</li>
+<li>Added 'Target', 'Visibility', 'Toggle Combat State' and 'Set Condition' options to the 'Token' action (under 'on click')</li>
+<li>Added 'Zoom In' and 'Zoom Out' functions to 'Move' action</li>
+<li>Added ability to load icons from web sources</li>
+<li>Added support for localization (module only)</li>
+<li>Fixed issue where SD buttons would not load properly if Foundry was initialized before the SD, or if Foundry was refreshed</li>
+<li>Long words now split up onto multiple lines</li>
+</ul>
+<b>Compatible server app and SD plugin:</b><br>
+Server v0.2.4 (no change)<br>
+SD plugin v0.7.2<br>
+
 ### v0.8.4 - 11-11-2020
 <ul>
 <li>In 'Other' action, 'Control Button' mode, add 'Displayed Controls' option</li>
@@ -9,7 +27,7 @@
 <li>Newly created actions now show the correct options at the start</li>
 </ul>
 <b>Compatible server app and SD plugin:</b><br>
-Server v0.2.4<br>
+Server v0.2.4 (no change)<br>
 SD plugin v0.7.1<br>
 
 ### v0.8.3 - 10-11-2020